Various changes
- Rename PlanetaryClusteringV2 as KMeansClustering to be more general
- Replace references to PlanetaryClusteringV2 with references to KMeansClustering
- When writing out the cluster labels, remove the ".0" that is added to the cluster number, since DATAVIEW_MathMatrix is all doubles
- Change the output structure so that there are 2 output files:
   1) Output from ClusterResult writer prints the output data with IDs as the first column (corresponding to the order the original data set was in), with the final column as the assigned cluster.
   2) Output from AllClusterSSE gives the SSE of each individual cluster and the total, all in one output file

+++ DATAVIEW/src/OneClusterSSE.java
@@ -15,13 +15,16 @@ public class OneClusterSSE extends Task {
     @Override
     public void run() {
         DATAVIEW_MathMatrix matrix = (DATAVIEW_MathMatrix) ins[0].read();
-        // drop the final column from the data, which is the column that gives the column number
-        DATAVIEW_MathMatrix data = DATAVIEW_MathMatrix.dropColumn(matrix.getNumOfColumns() - 1, matrix);
-        DATAVIEW_MathVector centroid = (DATAVIEW_MathVector) ins[1].read();
+        double sse = 0.0;
 
-        double sse = 0;
-        for (int row = 0; row < data.getNumOfRows(); row++) {
-            sse += squaredEuclideanDist(data.getRow(row), centroid);
+        if (matrix.getNumOfRows() != 0) {  // there could be 0 points assigned to this cluster
+            // drop the final column from the data, which is the column that gives the column number
+            DATAVIEW_MathMatrix data = DATAVIEW_MathMatrix.dropColumn(matrix.getNumOfColumns() - 1, matrix);
+            DATAVIEW_MathVector centroid = (DATAVIEW_MathVector) ins[1].read();
+
+            for (int row = 0; row < data.getNumOfRows(); row++) {
+                sse += squaredEuclideanDist(data.getRow(row), centroid);
+            }
         }
         outs[0].write(sse);
     }
